creo que funciono el revert de revert
