+++ public/javascripts/controllers/NavCtrl.js
@@ -1,0 +1,9 @@
+app.controller('NavCtrl', [
+	'$scope',
+	'auth',
+	function($scope, auth){
+	  $scope.isLoggedIn = auth.isLoggedIn;
+	  $scope.currentUser = auth.currentUser;
+	  $scope.logOut = auth.logOut;
+	}
+]);
