+++ views/index.ejs
@@ -4,6 +4,14 @@
     <link href="http://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/css/bootstrap.min.css" rel="stylesheet">
     <script src="http://ajax.googleapis.com/ajax/libs/angularjs/1.5.2/angular.min.js"></script>
     <script src="/javascripts/angularApp.js"></script>
+    <script src="/javascripts/services/ideas.js"></script>
+    <script src="/javascripts/services/auth.js"></script>
+    <script src="/javascripts/services/activities.js"></script>
+    <script src="/javascripts/controllers/NavCtrl.js"></script>
+    <script src="/javascripts/controllers/AuthCtrl.js"></script>
+    <script src="/javascripts/controllers/MainCtrl.js"></script>
+    <script src="/javascripts/controllers/IdeasCtrl.js"></script>
+    <script src="/javascripts/controllers/PendingIdeasCtrl.js"></script>
     <script src="http://cdnjs.cloudflare.com/ajax/libs/angular-ui-router/0.2.10/angular-ui-router.js"></script>
     
     <script src="moment.min.js"></script>
